removed modal context and replaced with consistent usedialogstate hook for all dialogs
