+++ src/components/forms/search/SearchSettings.jsx
@@ -3,7 +3,7 @@ import React, { useState } from 'react';
 import { Paper, Box, Container, Typography } from '@mui/material';
 import SearchFormB from './SearchFormB'; // Ensure path is correct
 import { useMode } from '../../../context';
-import { SearchSettingsBox } from '../../../pages/pageStyles/StyledComponents';
+import { SearchSettingsBox } from '../../../layout/REUSABLE_STYLED_COMPONENTS/ReusableStyledComponents';
 
 export const commonPaperStyles = (theme) => ({
   padding: theme.spacing(3),
