+++ src/context/index.js
@@ -3,7 +3,6 @@ export { useCartStore } from './MAIN_CONTEXT/CartContext/CartContext';
 export { useDeckStore } from './MAIN_CONTEXT/DeckContext/DeckContext';
 export { useCardStore } from './MAIN_CONTEXT/CardContext/CardContext';
 export { useCollectionStore } from './MAIN_CONTEXT/CollectionContext/CollectionContext';
-export { useModalContext } from './UTILITIES_CONTEXT/ModalContext/ModalContext';
 export { useUserContext } from './MAIN_CONTEXT/UserContext/UserContext';
 export { useSidebarContext } from './UTILITIES_CONTEXT/SideBarContext/SideBarProvider';
 export { useAppContext } from './MISC_CONTEXT/AppContext/AppContextProvider';
@@ -20,7 +19,6 @@ export { CardProvider } from './MAIN_CONTEXT/CardContext/CardContext';
 export { CollectionProvider } from './MAIN_CONTEXT/CollectionContext/CollectionContext';
 export { AppContextProvider } from './MISC_CONTEXT/AppContext/AppContextProvider';
 export { UserProvider } from './MAIN_CONTEXT/UserContext/UserContext';
-export { ModalProvider } from './UTILITIES_CONTEXT/ModalContext/ModalContext';
 export { ColorModeProvider } from './UTILITIES_CONTEXT/ColorModeContext/ColorModeProvider';
 export { SidebarProvider } from './UTILITIES_CONTEXT/SideBarContext/SideBarProvider';
 export { ConfiguratorProvider } from './UTILITIES_CONTEXT/ConfiguratorContext/ConfiguratorContext';
