+++ src/layout/REUSABLE_COMPONENTS/layout-utils/PageLayout.jsx
@@ -14,7 +14,11 @@ function PageLayout({ background, backCol, children }) {
       height="100%"
       minHeight="100vh"
       maxWidth="100%"
-      // bgColor={backCol ? '#3D3D3D' : 'transparent'}
+      // bgColor={
+      //   backCol
+      //     ? 'linear-gradient(90deg, rgba(13, 93, 150, 0.3) 0%, rgba(160, 214, 186, 0.3) 100%)'
+      //     : 'transparent'
+      // }
       sx={{ overflowX: 'hidden', m: 0, p: 0 }} // Ensure no margins or paddings
     >
       <Grid
