+++ src/layout/REUSABLE_STYLED_COMPONENTS/ReusableStyledComponents.jsx
@@ -7,6 +7,10 @@ import {
   Dialog,
   DialogActions,
   DialogContent,
+  FormControl,
+  InputLabel,
+  List,
+  ListItem,
   Paper,
   Popper,
   TextField,
@@ -50,7 +54,8 @@ export const StyledContainerBox = styled(Box)(({ theme }) => ({
   marginTop: theme.spacing(2),
   padding: theme.spacing(3),
   borderRadius: theme.shape.borderRadius,
-  background: 'white',
+  background:
+    'linear-gradient(90deg, rgba(13, 93, 150, 0.3) 0%, rgba(160, 214, 186, 0.3) 100%)',
   // background: theme.palette.greenAccent.crystalGreen,
   boxShadow: theme.shadows[10],
   marginBottom: theme.spacing(4),
@@ -74,6 +79,9 @@ export const StyledPaper = styled(Box)(({ theme }) => ({
   justifyContent: 'center',
   mx: 'auto',
   padding: '1rem',
+  // background: 'tra|nsparent',
+  // background:
+  //   'linear-gradient(90deg, rgba(13, 93, 150, 0.3) 0%, rgba(160, 214, 186, 0.3) 100%)',
   background: theme.palette.grey.clearGrey,
   // maxWidth: '1200px',
   borderRadius: '8px',
@@ -283,7 +291,11 @@ export const CardDetailContainer = styled(Box)(({ theme }) => ({
     boxShadow: `0 6px 12px 0 ${theme.palette.shadow}`,
   },
 }));
-
+export const SearchSettingsBox = styled(Box)(({ theme }) => ({
+  display: 'flex',
+  flexDirection: 'column',
+  gap: theme.spacing(2),
+}));
 export const AspectRatioBox = styled(Box)(({ theme }) => ({
   width: '100%', // Full width of the parent container
   position: 'relative',
@@ -325,4 +337,153 @@ export const StyledCard = styled(Card)(({ theme }) => ({
 export const QuantityLine = styled(Typography)(({ theme }) => ({
   fontSize: theme.typography.pxToRem(12),
   color: theme.palette.text.primary,
+}));
+// ! PORTFOLIO CHART STYLES
+export const ChartArea = styled(Container)(({ theme }) => ({
+  height: '100%',
+  width: '100%',
+  // minHeight: '500px',
+  position: 'relative',
+  flexGrow: 1,
+  padding: theme.spacing(2),
+  margin: theme.spacing(2),
+  display: 'flex',
+  flexDirection: 'column',
+  alignItems: 'center',
+  justifyContent: 'center',
+  border: `1px solid ${theme.palette.divider}`,
+  borderRadius: theme.shape.borderRadius,
+  background: theme.palette.grey.contrastText,
+}));
+export const SquareChartContainer = styled(Box)(({ theme }) => ({
+  position: 'relative',
+  width: '100%', // 100% of the parent width
+  paddingTop: '100%', // Maintain aspect ratio (1:1)
+  overflow: 'hidden',
+  '& > *': {
+    position: 'absolute',
+    top: 0,
+    left: 0,
+    right: 0,
+    bottom: 0,
+  },
+}));
+export const ChartPaper = styled(Paper)(({ theme }) => ({
+  borderRadius: theme.shape.borderRadius,
+  boxShadow: theme.shadows[5],
+  backgroundColor: theme.palette.greenAccent.contrastText,
+  color: theme.palette.text.secondary,
+  padding: theme.spacing(2),
+  display: 'flex',
+  flexDirection: 'column',
+  alignItems: 'center',
+  justifyContent: 'center',
+  width: '100%',
+  minHeight: '400px',
+  overflow: 'hidden',
+  margin: theme.spacing(2, 0),
+  flexGrow: 1,
+}));
+export const ResponsiveSquare = styled(Box)(({ theme }) => ({
+  width: '100%',
+  paddingTop: '100%',
+  backgroundColor: theme.palette.greenAccent.contrastText,
+  borderRadius: theme.shape.borderRadius,
+  boxShadow: theme.shadows[5],
+  display: 'flex',
+  flexGrow: 1,
+  alignItems: 'center',
+  justifyContent: 'center',
+}));
+// ! FORMS / INPUTS
+export const StyledFormControl = styled(FormControl)(({ theme }) => ({
+  margin: theme.spacing(1, 0),
+  backgroundColor: theme.palette.greenAccent.contrastText, // Adjusted for a slight contrast
+  borderRadius: theme.shape.borderRadius,
+  boxShadow: theme.shadows[1], // Subtle shadow for depth
+
+  '& .MuiFilledInput-root': {
+    borderRadius: theme.shape.borderRadius,
+    '&:hover': {
+      backgroundColor: theme.palette.action.hover,
+    },
+    '&.Mui-focused': {
+      backgroundColor: theme.palette.greenAccent.contrastText,
+      borderColor: theme.palette.greenAccent.light,
+    },
+  },
+}));
+export const StyledInputLabel = styled(InputLabel)(({ theme }) => ({
+  fontWeight: 'bold', // Making label text bold
+  color: theme.palette.grey.simpleGrey,
+}));
+// ! SELECT COLLECTION LIST
+export const StyledSkeletonCard = styled(Card)(({ theme }) => ({
+  // Use the same styles as in StyledCard
+  display: 'flex',
+  flexDirection: 'column',
+  // minWidth: '109px',
+  minWidth: '100%',
+  maxWidth: '100%',
+  width: '100%',
+  height: '100%',
+  // minHeight: '100%',
+  minHeight: 100,
+  alignItems: 'center',
+  flexGrow: 1,
+  // width: 'auto',
+  maxHeight: '14vh',
+  backgroundColor: theme.palette.greenAccent.contrastText,
+  borderRadius: theme.shape.borderRadius,
+  boxShadow: theme.shadows[5],
+  transition: 'transform 0.3s ease-in-out',
+  '&:hover': {
+    transform: 'scale(1.03)',
+  },
+}));
+export const ListItemSkeleton = styled(ListItem)(({ theme }) => ({
+  // margin: theme.spacing(1, 0),
+  // borderRadius: theme.shape.borderRadius,
+  margin: theme.spacing(1),
+  width: '100%',
+  minWidth: 200,
+  minHeight: 200,
+  maxHeight: 200,
+  flexGrow: 1,
+  transition: '0.3s',
+  boxShadow: '0 8px 40px -12px rgba(0,0,0,0.3)',
+  '&:hover': {
+    boxShadow: '0 16px 70px -12.125px rgba(0,0,0,0.3)',
+  },
+}));
+export const AspectRatioBoxSkeleton = styled(Box)(({ theme }) => ({
+  width: '100%',
+  position: 'relative',
+  paddingTop: '56.25%', // 16:9 aspect ratio
+  flexGrow: '1',
+}));
+export const PortfolioBoxA = styled(Box)(({ theme }) => ({
+  flexGrow: 1,
+  boxShadow: theme.shadows[5],
+  [theme.breakpoints.down('sm')]: {},
+}));
+export const CardUnorderedList = styled(List)(({ theme }) => ({
+  listStyleType: 'disc', // or 'circle' or 'square' for different bullet styles
+  paddingLeft: theme.spacing(4), // Adjust based on theme spacing for indentation
+  margin: 0, // Remove default margins
+}));
+export const CardListItem = styled(ListItem)(({ theme }) => ({
+  color: theme.palette.text.primary,
+  paddingBottom: theme.spacing(1), // Space between list items
+  textAlign: 'left', // Align text to the left
+  fontSize: '1rem', // Adjust font size as needed
+}));
+export const FeatureCard = styled(Card)(({ theme }) => ({
+  background: theme.palette.greenAccent.lightest,
+  boxShadow: theme.shadows[5],
+  transition: 'box-shadow 0.3s ease-in-out', // smooth transition for shadow
+  '&:hover': {
+    boxShadow: theme.shadows[15], // more prominent shadow on hover
+    transform: 'translateY(-3px)', // slight upward lift
+  },
 }));
