+++ src/layout/store/index.jsx
@@ -2,8 +2,8 @@ import React, { useState } from 'react';
 import { useMode } from '../../context';
 import MDBox from '../REUSABLE_COMPONENTS/MDBOX';
 import PageLayout from '../REUSABLE_COMPONENTS/layout-utils/PageLayout';
-import { PortfolioBoxA } from '../../pages/pageStyles/StyledComponents';
 import SearchComponent from '../../components/forms/search/SearchComponent';
+import { PortfolioBoxA } from '../REUSABLE_STYLED_COMPONENTS/ReusableStyledComponents';
 
 const StoreSearch = () => {
   const { theme } = useMode();
