+++ src/pages/DeckBuilderPage.js
@@ -1,30 +1,24 @@
 import React, { useEffect } from 'react';
 import GenericCardDialog from '../components/dialogs/GenericCardDialog';
 import DeckBuilder from '../layout/deck';
-import useLoadingAndModal from './pageStyles/useLoadingAndModal';
 import PageLayout from '../layout/REUSABLE_COMPONENTS/layout-utils/PageLayout';
 import useDeckManager from '../context/MAIN_CONTEXT/DeckContext/useDeckManager';
+import useDialogState from '../context/hooks/useDialogState';
 
 const DeckBuilderPage = () => {
-  const { closeModal, returnDisplay, isModalOpen, modalContent } =
-    useLoadingAndModal();
+  const { closeDialog, dialogState, data } = useDialogState();
   const { hasFetchedDecks, fetchDecks } = useDeckManager();
-  // useEffect(() => {
-  //   if (!hasFetchedDecks) {
-  //     fetchDecks();
-  //   }
-  // }, []);
-
   return (
     <PageLayout>
       <DeckBuilder />
-      {isModalOpen && (
+      {/* {dialogState.isCardDialogOpen && (
         <GenericCardDialog
-          open={isModalOpen}
+          open={dialogState.isCardDialogOpen}
           context={'Deck'}
-          card={modalContent}
+          onClose={() => closeDialog('isCardDialogOpen')}
+          card={data}
         />
-      )}
+      )} */}
     </PageLayout>
   );
 };
