--- src/pages/pageStyles/useLoadingAndModal.jsx
@@ -1,14 +1,0 @@
-// This is a conceptual example, adjust based on actual logic
-import { useModalContext } from '../../context';
-
-const useLoadingAndModal = () => {
-  const { isModalOpen, modalContent, closeModal } = useModalContext();
-
-  return {
-    isModalOpen,
-    modalContent,
-    closeModal,
-  };
-};
-
-export default useLoadingAndModal;
